Commit before putting refactor project on hold
This commit was initialy meant to sit on the BitcoiJ refactor branch,
but decided to put it in a separate repo. The message below was not
rephrased.

CreatePeerGroup2 is a relatively simple spec of a small part of the
process that decides how to source peers. DecidePeerSourcing was written
in the later days of the refactor and is much more complete. Neither of
these specs are meant to be published for general contributor (or even
reviewer) usage as is. They're rough drafts for my own development, but
since DecidePeerSourcing represents close to 50% of the work I've done
on this project, I'm commiting this into the WIP branch.